Relocate tests closer to query implementations
