--- graphql/general/src/queries/tests/mod.rs
@@ -1,4 +1,0 @@
-mod item_stats;
-mod items;
-mod master_lists;
-mod names;
